fucking predict surgery (#4040)
* predict the fucking start surgery verb

* replicate self surgery cvar

* goida

* FUCK IT PREDICT SURGERY ALTOGETHER

---------
