+++ Content.Shared/_Shitmed/Surgery/SharedSurgerySystem.Steps.cs
@@ -908,7 +908,7 @@ public abstract partial class SharedSurgerySystem
             locResult = Loc.GetString($"surgery-popup-step-{stepId}",
                 ("user", userName), ("target", targetName), ("part", part));
 
-        _popup.PopupEntity(locResult, user);
+        _popup.PopupPredicted(locResult, user, user);
         return true;
     }
 
@@ -1013,7 +1013,7 @@ public abstract partial class SharedSurgerySystem
         if (check.Invalid != StepInvalidReason.None)
         {
             if (doPopup && check.Popup != null)
-                _popup.PopupEntity(check.Popup, user, user, PopupType.SmallCaution);
+                _popup.PopupClient(check.Popup, user, user, PopupType.SmallCaution);
 
             reason = check.Invalid;
             return false;
